+++ src/Store.fs
@@ -7,6 +7,30 @@ module Store =
     let connect (configuration: ConnectionConfiguration) =
         new AerospikeClient(configuration.Host, configuration.Port)
 
+    let tryToConnect logAerospike configuration =
+        try
+            connect configuration |> Some
+        with
+        | :? AerospikeException as e ->
+            logAerospike <| sprintf "%A" e
+            None
+
+    let rec tryToConnectWithReconnects logAerospike attempts configuration =
+        let attemptLeft = (attempts - 1)
+        logAerospike <| sprintf "Try to connect - attempt left: %i" attemptLeft
+        if attemptLeft <= 0 then
+            failwithf "Connection to aerospike could not be estabilished"
+
+        match configuration |> tryToConnect logAerospike with
+        | Some aerospike -> aerospike
+        | None ->
+            logAerospike "Connection failed, waiting for 10s ..."
+            System.Threading.Thread.Sleep (10 * 1000)
+            tryToConnectWithReconnects logAerospike attemptLeft configuration
+
+    let connectWithReconnects logAerospike configuration =
+        configuration |> tryToConnectWithReconnects logAerospike 10
+
     let private createKey configuration (keyValue: string) =
         Key(configuration.Namespace, configuration.SetName, keyValue)